refactor: move profile editing to NavigationBar component
