+++ src/app/page.tsx
@@ -1,18 +1,14 @@
 "use client";
 import { useEffect, useState } from "react";
 import { useRouter } from "next/navigation";
-import { Container, Group, Button, Title, Box, Paper, rem, Text, Modal, TextInput, Card, Stack, Loader } from "@mantine/core";
+import { Container, Group, Button, Title, Box, Text, Modal, TextInput, Card, Stack, Loader } from "@mantine/core";
 import { showNotification } from "@mantine/notifications";
 import Link from "next/link";
 import { NavigationBar } from "@/components/NavigationBar";
-import { IconMaximize } from "@tabler/icons-react";
 
 export default function Home() {
   const router = useRouter();
   const [userName, setUserName] = useState<string | null>(null);
-  const [modalOpened, setModalOpened] = useState(false);
-  const [editName, setEditName] = useState("");
-  const [saving, setSaving] = useState(false);
 
   // Projects state
   const [projects, setProjects] = useState<any[]>([]);
@@ -20,19 +16,6 @@ export default function Home() {
   const [projectModal, setProjectModal] = useState(false);
   const [newProjectName, setNewProjectName] = useState("");
   const [creatingProject, setCreatingProject] = useState(false);
-
-  // Profile modal state for password change
-  const [currentPassword, setCurrentPassword] = useState("");
-  const [newPassword, setNewPassword] = useState("");
-  const [confirmPassword, setConfirmPassword] = useState("");
-  const [changingPassword, setChangingPassword] = useState(false);
-
-  // Delete account state
-  const [deletingAccount, setDeletingAccount] = useState(false);
-  const [deleteConfirmOpen, setDeleteConfirmOpen] = useState(false);
-
-  // Debug: log on each render
-  console.log("Rendering Home, userName:", userName);
 
   useEffect(() => {
     const token = localStorage.getItem("token");
@@ -81,49 +64,6 @@ export default function Home() {
     router.replace("/login");
   };
 
-  const handleNameClick = () => {
-    setEditName(userName || "");
-    setModalOpened(true);
-  };
-
-  const handleSave = async () => {
-    setSaving(true);
-    try {
-      const userEmail = localStorage.getItem("user:username") || "";
-      if (!userEmail) throw new Error("User email not found in localStorage");
-      const res = await fetch(
-        `http://localhost:3333/users?mode=volatile&key=${encodeURIComponent(userEmail + ":username")}`,
-        {
-          method: "POST",
-          body: editName,
-        }
-      );
-      if (!res.ok) throw new Error((await res.text()) || "Failed to update name");
-      const user = localStorage.getItem("user");
-      if (user) {
-        const parsed = JSON.parse(user);
-        parsed.name = editName;
-        localStorage.setItem("user", JSON.stringify(parsed));
-      }
-      setUserName(editName);
-      console.log("Updated userName to:", editName);
-      setModalOpened(false);
-      showNotification({
-        title: "Success",
-        message: "Name updated successfully!",
-        color: "green",
-      });
-    } catch (err: any) {
-      showNotification({
-        title: "Error",
-        message: err.message || "Failed to update name.",
-        color: "red",
-      });
-    } finally {
-      setSaving(false);
-    }
-  };
-
   const handleCreateProject = async () => {
     setCreatingProject(true);
     try {
@@ -159,99 +99,6 @@ export default function Home() {
     }
   };
 
-  const handleChangePassword = async () => {
-    setChangingPassword(true);
-    try {
-      if (!currentPassword || !newPassword || !confirmPassword) {
-        showNotification({ title: "Error", message: "All fields are required.", color: "red" });
-        return;
-      }
-      if (newPassword !== confirmPassword) {
-        showNotification({ title: "Error", message: "New passwords do not match.", color: "red" });
-        return;
-      }
-      const userEmail = localStorage.getItem("user:username") || "";
-      if (!userEmail) throw new Error("User email not found in localStorage");
-      // Fetch encrypted password from Civil Memory
-      const res = await fetch(
-        `http://localhost:3333/users?mode=volatile&key=${encodeURIComponent(userEmail)}`
-      );
-      if (!res.ok) throw new Error("Failed to fetch current password");
-      const encryptedPassword = await res.text();
-      // Extract salt from stored encrypted password
-      const saltB64 = extractSaltFromEncrypted(encryptedPassword);
-      // Encrypt entered current password with extracted salt
-      const encryptedInput = await encryptPassword(currentPassword, saltB64);
-      if (encryptedInput !== encryptedPassword) {
-        showNotification({ title: "Error", message: "Current password is incorrect.", color: "red" });
-        return;
-      }
-      // Encrypt new password with a new salt
-      const newEncrypted = await encryptPassword(newPassword);
-      // Store new encrypted password
-      const resSet = await fetch(
-        `http://localhost:3333/users?mode=volatile&key=${encodeURIComponent(userEmail)}`,
-        {
-          method: "POST",
-          body: newEncrypted,
-        }
-      );
-      if (!resSet.ok) throw new Error("Failed to update password");
-      showNotification({ title: "Success", message: "Password changed successfully!", color: "green" });
-      setCurrentPassword("");
-      setNewPassword("");
-      setConfirmPassword("");
-    } catch (err: any) {
-      showNotification({ title: "Error", message: err.message || "Failed to change password.", color: "red" });
-    } finally {
-      setChangingPassword(false);
-    }
-  };
-
-  const handleDeleteAccount = async () => {
-    setDeletingAccount(true);
-    try {
-      const userEmail = localStorage.getItem("user:username") || "";
-      if (!userEmail) throw new Error("User email not found in localStorage");
-      // Delete user password
-      await fetch(`http://localhost:3333/users?mode=volatile&key=${encodeURIComponent(userEmail)}`, { method: "DELETE" });
-      // Delete user name
-      await fetch(`http://localhost:3333/users?mode=volatile&key=${encodeURIComponent(userEmail + ':username')}`, { method: "DELETE" });
-      // Remove user from all projects
-      const res = await fetch("http://localhost:3333/projects?mode=volatile&key=projects");
-      if (res.ok) {
-        const projects = await res.json();
-        let changed = false;
-        const updatedProjects = Array.isArray(projects)
-          ? projects.map((p) => {
-            if (Array.isArray(p.members) && p.members.includes(userEmail)) {
-              changed = true;
-              return { ...p, members: p.members.filter((m: string) => m !== userEmail) };
-            }
-            return p;
-          })
-          : [];
-        if (changed) {
-          await fetch("http://localhost:3333/projects?mode=volatile&key=projects", {
-            method: "POST",
-            body: JSON.stringify(updatedProjects),
-          });
-        }
-      }
-      // Log out
-      localStorage.removeItem("token");
-      localStorage.removeItem("user");
-      localStorage.removeItem("user:username");
-      showNotification({ title: "Account Deleted", message: "Your account and data have been deleted.", color: "green" });
-      router.replace("/login");
-    } catch (err: any) {
-      showNotification({ title: "Error", message: err.message || "Failed to delete account.", color: "red" });
-    } finally {
-      setDeletingAccount(false);
-      setDeleteConfirmOpen(false);
-    }
-  };
-
   return (
     <Box style={{ minHeight: "100vh", background: "linear-gradient(135deg, #f8fafc 0%, #e0e7ff 100%)" }}>
       <NavigationBar userName={userName} onLogout={handleLogout} />
@@ -266,53 +113,6 @@ export default function Home() {
         <Button onClick={handleCreateProject} loading={creatingProject} fullWidth>
           Create
         </Button>
-      </Modal>
-      <Modal opened={modalOpened} onClose={() => setModalOpened(false)} title="Edit Profile" centered>
-        <TextInput
-          label="Name"
-          value={editName}
-          onChange={(e) => setEditName(e.currentTarget.value)}
-          mb="md"
-        />
-        <Button onClick={() => { console.log("Save button clicked"); handleSave(); }} loading={saving} fullWidth mb="md">
-          Save
-        </Button>
-        <Title order={5} mt="md" mb="xs">Change Password</Title>
-        <TextInput
-          label="Current Password"
-          type="password"
-          value={currentPassword}
-          onChange={(e) => setCurrentPassword(e.currentTarget.value)}
-          mb="xs"
-        />
-        <TextInput
-          label="New Password"
-          type="password"
-          value={newPassword}
-          onChange={(e) => setNewPassword(e.currentTarget.value)}
-          mb="xs"
-        />
-        <TextInput
-          label="Confirm New Password"
-          type="password"
-          value={confirmPassword}
-          onChange={(e) => setConfirmPassword(e.currentTarget.value)}
-          mb="md"
-        />
-        <Button onClick={handleChangePassword} loading={changingPassword} fullWidth color="violet" mb="md">
-          Change Password
-        </Button>
-        <Button color="red" variant="outline" fullWidth onClick={() => setDeleteConfirmOpen(true)} loading={deletingAccount}>
-          Delete Account
-        </Button>
-      </Modal>
-      {/* Delete confirmation modal */}
-      <Modal opened={deleteConfirmOpen} onClose={() => setDeleteConfirmOpen(false)} title="Confirm Account Deletion" centered>
-        <Text c="red" mb="md">Are you sure you want to permanently delete your account and all associated data? This action cannot be undone.</Text>
-        <Group grow>
-          <Button variant="default" onClick={() => setDeleteConfirmOpen(false)} disabled={deletingAccount}>Cancel</Button>
-          <Button color="red" onClick={handleDeleteAccount} loading={deletingAccount}>Delete</Button>
-        </Group>
       </Modal>
       <Container size="lg" mt={40}>
         <Title order={2} mb="md">Welcome to SparkPad!</Title>
@@ -337,71 +137,4 @@ export default function Home() {
       </Container>
     </Box>
   );
-}
-
-async function encryptPassword(password: string, saltB64?: string): Promise<string> {
-  if (!saltB64) {
-    // Generate new salt
-    const salt = window.crypto.getRandomValues(new Uint8Array(16));
-    const enc = new TextEncoder();
-    const passwordKey = await window.crypto.subtle.importKey(
-      "raw",
-      enc.encode(password),
-      { name: "PBKDF2" },
-      false,
-      ["deriveBits", "deriveKey"]
-    );
-    const derivedKey = await window.crypto.subtle.deriveKey(
-      {
-        name: "PBKDF2",
-        salt,
-        iterations: 100000,
-        hash: "SHA-256",
-      },
-      passwordKey,
-      { name: "AES-GCM", length: 256 },
-      true,
-      ["encrypt", "decrypt"]
-    );
-    const rawKey = await window.crypto.subtle.exportKey("raw", derivedKey);
-    const combined = new Uint8Array(salt.length + rawKey.byteLength);
-    combined.set(salt, 0);
-    combined.set(new Uint8Array(rawKey), salt.length);
-    return btoa(String.fromCharCode(...combined));
-  } else {
-    // Use provided salt
-    const salt = Uint8Array.from(atob(saltB64), c => c.charCodeAt(0));
-    const enc = new TextEncoder();
-    const passwordKey = await window.crypto.subtle.importKey(
-      "raw",
-      enc.encode(password),
-      { name: "PBKDF2" },
-      false,
-      ["deriveBits", "deriveKey"]
-    );
-    const derivedKey = await window.crypto.subtle.deriveKey(
-      {
-        name: "PBKDF2",
-        salt,
-        iterations: 100000,
-        hash: "SHA-256",
-      },
-      passwordKey,
-      { name: "AES-GCM", length: 256 },
-      true,
-      ["encrypt", "decrypt"]
-    );
-    const rawKey = await window.crypto.subtle.exportKey("raw", derivedKey);
-    const combined = new Uint8Array(salt.length + rawKey.byteLength);
-    combined.set(salt, 0);
-    combined.set(new Uint8Array(rawKey), salt.length);
-    return btoa(String.fromCharCode(...combined));
-  }
-}
-
-function extractSaltFromEncrypted(encrypted: string): string {
-  // The salt is the first 16 bytes (base64-encoded)
-  const bytes = Uint8Array.from(atob(encrypted), c => c.charCodeAt(0));
-  const salt = bytes.slice(0, 16);
-  return btoa(String.fromCharCode(...salt));
 }
